entrypoint/requirements.txt
backports.ssl-match-hostname==3.5.0.1
click==6.6
docker-py==1.9.0
Flask==0.11.1
Flask-Cors==3.0.1
gevent==1.1.2
greenlet==0.4.10
gunicorn==19.6.0
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
requests==2.11.1
six==1.10.0
websocket-client==0.37.0
Werkzeug==0.11.11
